reorganizacion de la estructura
