+++ Dutify/src/components/searchBar/searchBarStyle.css
@@ -8,7 +8,7 @@
 
 .search-btn {
   right: 3px;
-  top: 5%;
+  top: 6%;
   background-color: #655be6;
   border-radius: 50%;
   border: 1px solid black;
