+++ Dutify/src/components/themeSwitch/themeSwitch.jsx
@@ -1,0 +1,13 @@
+import "./themeSwitchStyle.css"
+
+function ThemeSwitch() {
+  return (
+    <button className="theme-toogle">
+        <svg>
+            
+        </svg>
+    </button>
+  );
+}
+
+export default ThemeSwitch;
