+++ Dutify/src/components/themeSwitch/themeSwitchStyle.css
@@ -1,0 +1,4 @@
+.theme-toogle{
+  background-color: transparent;
+  border: 0;
+}
