+++ Dutify/src/index.css
@@ -1,5 +1,5 @@
 .topBar{
-  height: 5.156em;
+  height: 6vh;
   background-color: #4E4848;
 }
